requirements.txt
mysql-connector-python
pandas
PyMySQL
pyodbc